initial id support if it is enabled in conf #81

+++ src/main/kotlin/cs/ut/ui/controllers/modal/ParameterModalController.kt
@@ -94,10 +94,12 @@ class ParameterModalController : GenericAutowireComposer<Component>(), Redirecta
         grid.generate(cols)
 
         cancelBtn.addEventListener(Events.ON_CLICK, { _ ->
-            Files.delete(Paths.get(file.absolutePath))
-            Executions.getCurrent().desktop.components.firstOrNull { it.id == "upload" }?.let {
-                it as Button
-                it.isDisabled = false
+            if (!isRecreation) {
+                Files.delete(Paths.get(file.absolutePath))
+                Executions.getCurrent().desktop.components.firstOrNull { it.id == "upload" }?.let {
+                    it as Button
+                    it.isDisabled = false
+                }
             }
             modal.detach()
         })
